refactor: extract conversation actions and merged messages logic from chat component into custom hook
