+++ apps/chat/src/components/Chat/Chat.tsx
@@ -1,867 +1,16 @@
-import { FloatingOverlay } from '@floating-ui/react';
-import { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
-
 import { useTranslation } from 'next-i18next';
 
-import classNames from 'classnames';
-
 import { useChatSelectors } from '@/src/components/Chat/hooks/useChatSelectors';
-import { useChatViewAutoScroll } from '@/src/components/Chat/hooks/useChatViewAutoScroll';
-import { useChatViewSelectors } from '@/src/components/Chat/hooks/useChatViewSelectors';
 
-import { clearStateForMessages } from '@/src/utils/app/clear-messages-state';
-import { isSmallScreen } from '@/src/utils/app/mobile';
-
-import {
-  Conversation,
-  ConversationsTemporarySettings,
-  LikeState,
-  MergedMessages,
-  Message,
-  Replay,
-  Role,
-} from '@/src/types/chat';
-import { EntityType, UploadStatus } from '@/src/types/common';
+import { UploadStatus } from '@/src/types/common';
 import { Translation } from '@/src/types/translation';
 
-import { AddonsActions } from '@/src/store/addons/addons.reducers';
-import { ConversationsActions } from '@/src/store/conversations/conversations.reducers';
-import { useAppDispatch } from '@/src/store/hooks';
-import { ModelsActions } from '@/src/store/models/models.reducers';
-
-import { REPLAY_AS_IS_MODEL } from '@/src/constants/chat';
-import { DEFAULT_ASSISTANT_SUBMODEL_ID } from '@/src/constants/default-ui-settings';
+import { ChatView } from '@/src/components/Chat/ChatView';
 
 import Loader from '../Common/Loader';
 import { NotFoundEntity } from '../Common/NotFoundEntity';
-import { ChatCompareRotate } from './ChatCompareRotate';
-import { ChatCompareSelect } from './ChatCompareSelect';
-import ChatExternalControls from './ChatExternalControls';
-import { ChatHeader } from './ChatHeader';
-import { ChatInput } from './ChatInput/ChatInput';
 import { ChatInputFooter } from './ChatInput/ChatInputFooter';
-import { ChatSettings } from './ChatSettings';
-import { ChatSettingsEmpty } from './ChatSettingsEmpty';
-import { ErrorMessageDiv } from './ErrorMessageDiv';
-import { MemoizedChatMessage } from './MemoizedChatMessage';
-import { NotAllowedModel } from './NotAllowedModel';
-import { PlaybackControls } from './Playback/PlaybackControls';
-import { PublicationControls } from './Publish/PublicationChatControls';
 import { PublicationHandler } from './Publish/PublicationHandler';
-import { StartReplayButton } from './StartReplayButton';
-
-import { Feature } from '@epam/ai-dial-shared';
-
-export const ChatView = memo(() => {
-  const dispatch = useAppDispatch();
-
-  const {
-    appName,
-    models,
-    modelsMap,
-    modelError,
-    isModelsLoaded,
-    addons,
-    addonsMap,
-    isCompareMode,
-    selectedConversationsIds,
-    selectedConversations,
-    messageIsStreaming,
-    conversations,
-    prompts,
-    enabledFeatures,
-    isReplay,
-    isReplayPaused,
-    isReplayRequiresVariables,
-    isExternal,
-    isPlayback,
-    isAnyMenuOpen,
-    isIsolatedView,
-  } = useChatViewSelectors();
-
-  const [mergedMessages, setMergedMessages] = useState<MergedMessages[]>([]);
-  const [isShowChatSettings, setIsShowChatSettings] = useState(false);
-  const [isLastMessageError, setIsLastMessageError] = useState(false);
-  const [prevSelectedIds, setPrevSelectedIds] = useState<string[]>([]);
-  const [inputHeight, setInputHeight] = useState<number>(142);
-  const [notAllowedType, setNotAllowedType] = useState<EntityType | null>(null);
-
-  const selectedConversationsTemporarySettings = useRef<
-    Record<string, ConversationsTemporarySettings>
-  >({});
-  const chatContainerRef = useRef<HTMLDivElement | null>(null);
-  const textareaRef = useRef<HTMLTextAreaElement | null>(null);
-  const nextMessageBoxRef = useRef<HTMLDivElement | null>(null);
-  const chatMessagesRef = useRef<HTMLDivElement | null>(null);
-
-  const {
-    showScrollDownButton,
-    setShowScrollDownButton,
-    handleScroll,
-    handleScrollDown,
-    setAutoScroll,
-    scrollToContainerHeight,
-  } = useChatViewAutoScroll(
-    chatContainerRef,
-    chatMessagesRef,
-    mergedMessages.length,
-    messageIsStreaming,
-  );
-
-  const showReplayControls = useMemo(() => {
-    return (
-      isReplay &&
-      !messageIsStreaming &&
-      (isReplayPaused || !!isReplayRequiresVariables)
-    );
-  }, [isReplay, isReplayPaused, isReplayRequiresVariables, messageIsStreaming]);
-
-  const isNotEmptyConversations =
-    isReplayRequiresVariables ||
-    selectedConversations.some((conv) => conv.messages.length > 0);
-
-  useEffect(() => {
-    const isNotAllowedModel =
-      isModelsLoaded &&
-      (models.length === 0 ||
-        selectedConversations.some((conv) => {
-          if (
-            conv.replay &&
-            conv.replay.isReplay &&
-            conv.replay.replayAsIs &&
-            conv.replay.replayUserMessagesStack &&
-            conv.replay.replayUserMessagesStack[0].model
-          ) {
-            return conv.replay.replayUserMessagesStack.some(
-              (message) =>
-                message.role === Role.User &&
-                message.model?.id &&
-                !modelsMap[message.model.id],
-            );
-          }
-
-          const model = modelsMap[conv.model.id];
-
-          return (
-            !model ||
-            (model.type === EntityType.Assistant &&
-              conv.assistantModelId &&
-              !modelsMap[conv.assistantModelId])
-          );
-        }));
-    if (isNotAllowedModel) {
-      setNotAllowedType(EntityType.Model);
-    } else if (
-      selectedConversations.some((conversation) =>
-        conversation.selectedAddons.some((addonId) => !addonsMap[addonId]),
-      )
-    ) {
-      setNotAllowedType(EntityType.Addon);
-    } else {
-      setNotAllowedType(null);
-    }
-  }, [selectedConversations, models, isModelsLoaded, addonsMap, modelsMap]);
-
-  const onLikeHandler = useCallback(
-    (index: number, conversation: Conversation) => (rate: LikeState) => {
-      dispatch(
-        ConversationsActions.rateMessage({
-          conversationId: conversation.id,
-          messageIndex: index,
-          rate,
-        }),
-      );
-    },
-    [dispatch],
-  );
-
-  useEffect(() => {
-    const lastMergedMessages = mergedMessages.length
-      ? mergedMessages[mergedMessages.length - 1]
-      : [];
-
-    const isErrorInSomeLastMessage = lastMergedMessages.some(
-      (mergedStr: [Conversation, Message, number]) =>
-        !!mergedStr[1].errorMessage,
-    );
-    setIsLastMessageError(isErrorInSomeLastMessage);
-  }, [mergedMessages]);
-
-  useEffect(() => {
-    const handleResize = () => {
-      if (
-        chatMessagesRef.current &&
-        !messageIsStreaming &&
-        mergedMessages.length
-      ) {
-        handleScroll();
-      }
-    };
-
-    const resizeObserver = new ResizeObserver(handleResize);
-
-    if (chatMessagesRef.current) {
-      resizeObserver.observe(chatMessagesRef.current);
-    }
-
-    return () => {
-      resizeObserver.disconnect();
-    };
-  }, [handleScroll, mergedMessages.length, messageIsStreaming]);
-
-  useEffect(() => {
-    if (selectedConversations.length > 0) {
-      const mergedMessages: MergedMessages[] = [];
-      const firstConversationMessages =
-        selectedConversations[0].messages.filter((m) => m.role !== Role.System);
-      for (let i = 0; i < firstConversationMessages.length; i++) {
-        mergedMessages.push(
-          selectedConversations.map((conv) => [
-            conv,
-            conv.messages.filter((m) => m.role !== Role.System)[i] || {
-              role: Role.Assistant,
-              content: '',
-            },
-            i,
-          ]),
-        );
-      }
-
-      setMergedMessages(mergedMessages);
-    }
-
-    if (
-      selectedConversations.every(
-        (conv) => !conv.messages.find((m) => m.role !== Role.Assistant),
-      )
-    ) {
-      setShowScrollDownButton(false);
-    } else {
-      handleScroll();
-    }
-  }, [handleScroll, selectedConversations, setShowScrollDownButton]);
-
-  const handleClearConversation = useCallback(
-    (conversation: Conversation) => {
-      if (conversation) {
-        dispatch(
-          ConversationsActions.updateConversation({
-            id: conversation.id,
-            values: {
-              messages: [],
-            },
-          }),
-        );
-      }
-    },
-    [dispatch],
-  );
-
-  const applySelectedModel = useCallback(
-    (
-      conversation: Conversation,
-      modelId: string | undefined,
-    ): Partial<Conversation> => {
-      if (modelId === REPLAY_AS_IS_MODEL && conversation.replay) {
-        return {
-          replay: {
-            ...conversation.replay,
-            replayAsIs: true,
-          },
-        };
-      }
-      const newAiEntity = modelId ? modelsMap[modelId] : undefined;
-      if (!modelId || !newAiEntity) {
-        return {};
-      }
-
-      const updatedReplay: Replay | undefined = !conversation.replay?.isReplay
-        ? conversation.replay
-        : {
-            ...conversation.replay,
-            replayAsIs: false,
-          };
-      const updatedAddons =
-        conversation.replay &&
-        conversation.replay.isReplay &&
-        conversation.replay.replayAsIs &&
-        !updatedReplay?.replayAsIs
-          ? conversation.selectedAddons.filter((addonId) => addonsMap[addonId])
-          : conversation.selectedAddons;
-
-      return {
-        model: { id: newAiEntity.reference },
-        assistantModelId:
-          newAiEntity.type === EntityType.Assistant
-            ? DEFAULT_ASSISTANT_SUBMODEL_ID
-            : undefined,
-        replay: updatedReplay,
-        selectedAddons: updatedAddons,
-      };
-    },
-    [addonsMap, modelsMap],
-  );
-
-  const handleSelectModel = useCallback(
-    (conversation: Conversation, modelId: string) => {
-      const newAiEntity = modelsMap[modelId];
-      if (!newAiEntity && modelId !== REPLAY_AS_IS_MODEL) {
-        return;
-      }
-
-      dispatch(
-        ConversationsActions.updateConversation({
-          id: conversation.id,
-          values: {
-            ...applySelectedModel(conversation, modelId),
-          },
-        }),
-      );
-    },
-    [applySelectedModel, dispatch, modelsMap],
-  );
-
-  const handleSelectAssistantSubModel = useCallback(
-    (conversation: Conversation, modelId: string) => {
-      dispatch(
-        ConversationsActions.updateConversation({
-          id: conversation.id,
-          values: { assistantModelId: modelId },
-        }),
-      );
-    },
-    [dispatch],
-  );
-
-  useEffect(() => {
-    if (!selectedConversationsIds.some((id) => prevSelectedIds.includes(id))) {
-      setAutoScroll(true);
-      setPrevSelectedIds(selectedConversationsIds);
-      setIsShowChatSettings(false);
-    }
-  }, [prevSelectedIds, selectedConversationsIds, setAutoScroll]);
-
-  const handleOnChangeAddon = useCallback(
-    (conversation: Conversation, addonId: string) => {
-      const isAddonInConversation = conversation.selectedAddons.some(
-        (id) => id === addonId,
-      );
-      if (isAddonInConversation) {
-        const filteredAddons = conversation.selectedAddons.filter(
-          (id) => id !== addonId,
-        );
-        dispatch(
-          ConversationsActions.updateConversation({
-            id: conversation.id,
-            values: { selectedAddons: filteredAddons },
-          }),
-        );
-      } else {
-        dispatch(
-          ConversationsActions.updateConversation({
-            id: conversation.id,
-            values: {
-              selectedAddons: conversation.selectedAddons.concat(addonId),
-            },
-          }),
-        );
-      }
-    },
-    [dispatch],
-  );
-
-  const handleOnApplyAddons = useCallback(
-    (conversation: Conversation, addonIds: string[]) => {
-      dispatch(
-        ConversationsActions.updateConversation({
-          id: conversation.id,
-          values: {
-            selectedAddons: addonIds.filter((addonId) => addonsMap[addonId]),
-          },
-        }),
-      );
-    },
-    [addonsMap, dispatch],
-  );
-
-  const handleChangePrompt = useCallback(
-    (conversation: Conversation, prompt: string) => {
-      dispatch(
-        ConversationsActions.updateConversation({
-          id: conversation.id,
-          values: { prompt },
-        }),
-      );
-    },
-    [dispatch],
-  );
-
-  const handleChangeTemperature = useCallback(
-    (conversation: Conversation, temperature: number) => {
-      dispatch(
-        ConversationsActions.updateConversation({
-          id: conversation.id,
-          values: { temperature },
-        }),
-      );
-    },
-    [dispatch],
-  );
-
-  const handleDeleteMessage = useCallback(
-    (index: number, conv: Conversation) => {
-      let finalIndex = index;
-      if (conv.messages.at(0)?.role === Role.System) {
-        finalIndex += 1;
-      }
-      dispatch(ConversationsActions.deleteMessage({ index: finalIndex }));
-    },
-    [dispatch],
-  );
-
-  const onSendMessage = useCallback(
-    (message: Message) => {
-      dispatch(
-        ConversationsActions.sendMessages({
-          conversations: selectedConversations,
-          message,
-          deleteCount: 0,
-          activeReplayIndex: 0,
-        }),
-      );
-    },
-    [dispatch, selectedConversations],
-  );
-
-  const onRegenerateMessage = useCallback(() => {
-    const lastUserMessageIndex = selectedConversations[0].messages
-      .map((msg) => msg.role)
-      .lastIndexOf(Role.User);
-    dispatch(
-      ConversationsActions.sendMessages({
-        conversations: selectedConversations,
-        message: selectedConversations[0].messages[lastUserMessageIndex],
-        deleteCount:
-          selectedConversations[0].messages.length - lastUserMessageIndex,
-        activeReplayIndex: 0,
-      }),
-    );
-
-    scrollToContainerHeight('smooth');
-  }, [dispatch, scrollToContainerHeight, selectedConversations]);
-
-  const onEditMessage = useCallback(
-    (editedMessage: Message, index: number) => {
-      dispatch(ConversationsActions.stopStreamMessage());
-      dispatch(
-        ConversationsActions.sendMessages({
-          conversations: selectedConversations,
-          message: editedMessage,
-          deleteCount: mergedMessages.length - index,
-          activeReplayIndex: 0,
-        }),
-      );
-    },
-    [dispatch, mergedMessages.length, selectedConversations],
-  );
-
-  const handleApplyChatSettings = useCallback(() => {
-    selectedConversations.forEach((conversation) => {
-      const temporarySettings: ConversationsTemporarySettings | undefined =
-        selectedConversationsTemporarySettings.current[conversation.id];
-      if (temporarySettings) {
-        dispatch(
-          ConversationsActions.updateConversation({
-            id: conversation.id,
-            values: {
-              messages: clearStateForMessages(conversation.messages),
-              ...applySelectedModel(conversation, temporarySettings.modelId),
-              prompt: temporarySettings.prompt,
-              temperature: temporarySettings.temperature,
-              assistantModelId: temporarySettings.currentAssistentModelId,
-              selectedAddons: temporarySettings.addonsIds.filter(
-                (addonId) => addonsMap[addonId],
-              ),
-              isShared: temporarySettings.isShared,
-            },
-          }),
-        );
-      }
-    });
-  }, [selectedConversations, dispatch, applySelectedModel, addonsMap]);
-
-  const handleTemporarySettingsSave = useCallback(
-    (conversation: Conversation, args: ConversationsTemporarySettings) => {
-      selectedConversationsTemporarySettings.current[conversation.id] = args;
-    },
-    [],
-  );
-
-  const onChatInputResize = useCallback((inputHeight: number) => {
-    setInputHeight(inputHeight);
-  }, []);
-
-  const showLastMessageRegenerate =
-    !isReplay &&
-    !isPlayback &&
-    !isExternal &&
-    !messageIsStreaming &&
-    !isLastMessageError;
-  const showFloatingOverlay =
-    isSmallScreen() && isAnyMenuOpen && !isIsolatedView;
-
-  return (
-    <div
-      className="relative min-w-0 shrink grow basis-0 overflow-y-auto"
-      data-qa="chat"
-      id="chat"
-    >
-      {showFloatingOverlay && <FloatingOverlay className="z-30 bg-blackout" />}
-      {modelError ? (
-        <ErrorMessageDiv error={modelError} />
-      ) : (
-        <>
-          <div
-            className={classNames(
-              'flex size-full',
-              isCompareMode ? 'landscape:hidden' : 'hidden',
-            )}
-          >
-            <ChatCompareRotate />
-          </div>
-          <div
-            className={classNames(
-              'relative size-full',
-              isCompareMode && 'portrait:hidden',
-            )}
-          >
-            <div className="flex h-full">
-              <div
-                className={classNames(
-                  'flex h-full flex-col',
-                  isCompareMode && selectedConversations.length < 2
-                    ? 'w-[50%]'
-                    : 'w-full',
-                )}
-                data-qa={isCompareMode ? 'compare-mode' : 'chat-mode'}
-              >
-                <div className="flex h-full flex-col justify-between">
-                  <div className="flex w-full">
-                    {selectedConversations.map((conv) => (
-                      <div
-                        key={conv.id}
-                        className={classNames(
-                          isCompareMode && selectedConversations.length > 1
-                            ? 'w-[50%]'
-                            : 'w-full',
-                        )}
-                      >
-                        {conv.messages.length !== 0 &&
-                          enabledFeatures.has(Feature.TopSettings) && (
-                            <div className="z-10 flex flex-col">
-                              <ChatHeader
-                                conversation={conv}
-                                isCompareMode={isCompareMode}
-                                isShowChatInfo={enabledFeatures.has(
-                                  Feature.TopChatInfo,
-                                )}
-                                isShowClearConversation={
-                                  enabledFeatures.has(
-                                    Feature.TopClearConversation,
-                                  ) &&
-                                  !isPlayback &&
-                                  !isReplay &&
-                                  !messageIsStreaming &&
-                                  !isExternal
-                                }
-                                isShowModelSelect={
-                                  enabledFeatures.has(
-                                    Feature.TopChatModelSettings,
-                                  ) &&
-                                  !isPlayback &&
-                                  !isExternal
-                                }
-                                isShowSettings={isShowChatSettings}
-                                setShowSettings={(isShow) => {
-                                  if (isShow) {
-                                    dispatch(ModelsActions.getModels());
-                                    dispatch(AddonsActions.getAddons());
-                                  }
-                                  setIsShowChatSettings(isShow);
-                                }}
-                                selectedConversationIds={
-                                  selectedConversationsIds
-                                }
-                                onClearConversation={() =>
-                                  handleClearConversation(conv)
-                                }
-                                onUnselectConversation={(id) => {
-                                  dispatch(
-                                    ConversationsActions.unselectConversations({
-                                      conversationIds: [id],
-                                    }),
-                                  );
-                                }}
-                              />
-                            </div>
-                          )}
-                      </div>
-                    ))}
-                  </div>
-                  <div
-                    onScroll={() => {
-                      if (
-                        selectedConversations.some(
-                          (conv) =>
-                            !!conv.messages.find(
-                              (m) => m.role !== Role.Assistant,
-                            ),
-                        )
-                      ) {
-                        handleScroll();
-                      }
-                    }}
-                    ref={chatContainerRef}
-                    className="h-full overflow-x-hidden"
-                    data-qa="scrollable-area"
-                  >
-                    <div className="flex max-h-full w-full">
-                      {selectedConversations.map(
-                        (conv) =>
-                          conv.messages.length === 0 && (
-                            <div
-                              key={conv.id}
-                              className={classNames(
-                                'flex h-full flex-col justify-between',
-                                selectedConversations.length > 1
-                                  ? 'w-[50%]'
-                                  : 'w-full',
-                              )}
-                            >
-                              <div
-                                className="shrink-0"
-                                style={{
-                                  height: `calc(100% - ${inputHeight}px)`,
-                                }}
-                              >
-                                <ChatSettingsEmpty
-                                  conversation={conv}
-                                  isModels={models.length !== 0}
-                                  prompts={prompts}
-                                  isShowSettings={enabledFeatures.has(
-                                    Feature.EmptyChatSettings,
-                                  )}
-                                  onSelectModel={(modelId: string) =>
-                                    handleSelectModel(conv, modelId)
-                                  }
-                                  onSelectAssistantSubModel={(
-                                    modelId: string,
-                                  ) =>
-                                    handleSelectAssistantSubModel(conv, modelId)
-                                  }
-                                  onChangeAddon={(addonId: string) =>
-                                    handleOnChangeAddon(conv, addonId)
-                                  }
-                                  onChangePrompt={(prompt) =>
-                                    handleChangePrompt(conv, prompt)
-                                  }
-                                  onChangeTemperature={(temperature) =>
-                                    handleChangeTemperature(conv, temperature)
-                                  }
-                                  appName={appName}
-                                  onApplyAddons={handleOnApplyAddons}
-                                />
-                              </div>
-                            </div>
-                          ),
-                      )}
-                    </div>
-                    <div ref={chatMessagesRef}>
-                      {mergedMessages?.length > 0 && (
-                        <div className="flex flex-col" data-qa="chat-messages">
-                          {mergedMessages.map(
-                            (
-                              mergedStr: [Conversation, Message, number][],
-                              i: number,
-                            ) => (
-                              <div
-                                key={i}
-                                className="flex w-full"
-                                data-qa={
-                                  isCompareMode
-                                    ? 'compare-message-row'
-                                    : 'message-row'
-                                }
-                              >
-                                {mergedStr.map(
-                                  ([conv, message, index]: [
-                                    Conversation,
-                                    Message,
-                                    number,
-                                  ]) => (
-                                    <div
-                                      key={conv.id}
-                                      className={classNames(
-                                        isCompareMode &&
-                                          selectedConversations.length > 1
-                                          ? 'w-[50%]'
-                                          : 'w-full',
-                                      )}
-                                    >
-                                      <div className="size-full">
-                                        <MemoizedChatMessage
-                                          key={conv.id}
-                                          message={message}
-                                          messageIndex={index}
-                                          conversation={conv}
-                                          isLikesEnabled={enabledFeatures.has(
-                                            Feature.Likes,
-                                          )}
-                                          editDisabled={
-                                            !!notAllowedType ||
-                                            isExternal ||
-                                            isReplay ||
-                                            isPlayback
-                                          }
-                                          onEdit={onEditMessage}
-                                          onLike={onLikeHandler(index, conv)}
-                                          onDelete={() => {
-                                            handleDeleteMessage(index, conv);
-                                          }}
-                                          onRegenerate={
-                                            index ===
-                                              mergedMessages.length - 1 &&
-                                            showLastMessageRegenerate
-                                              ? onRegenerateMessage
-                                              : undefined
-                                          }
-                                          messagesLength={mergedMessages.length}
-                                        />
-                                      </div>
-                                    </div>
-                                  ),
-                                )}
-                              </div>
-                            ),
-                          )}
-                        </div>
-                      )}
-                    </div>
-                  </div>
-                  {!isPlayback && notAllowedType ? (
-                    <NotAllowedModel
-                      showScrollDownButton={showScrollDownButton}
-                      onScrollDownClick={handleScrollDown}
-                      type={notAllowedType}
-                    />
-                  ) : (
-                    <>
-                      {isExternal && selectedConversations.length === 1 && (
-                        <PublicationControls
-                          showScrollDownButton={showScrollDownButton}
-                          entity={selectedConversations[0]}
-                          onScrollDownClick={handleScrollDown}
-                          controlsClassNames="mx-2 mb-2 mt-5 w-full flex-row md:mx-4 md:mb-0 md:last:mb-6 lg:mx-auto lg:w-[768px] lg:max-w-3xl"
-                        />
-                      )}
-
-                      {!isPlayback && (
-                        <ChatInput
-                          showReplayControls={showReplayControls}
-                          textareaRef={textareaRef}
-                          showScrollDownButton={showScrollDownButton}
-                          onSend={onSendMessage}
-                          onScrollDownClick={handleScrollDown}
-                          onRegenerate={onRegenerateMessage}
-                          isLastMessageError={isLastMessageError}
-                          onStopConversation={() => {
-                            dispatch(ConversationsActions.stopStreamMessage());
-                          }}
-                          onResize={onChatInputResize}
-                          isShowInput={
-                            (!isReplay || isNotEmptyConversations) &&
-                            !isExternal
-                          }
-                        >
-                          {showReplayControls && !isNotEmptyConversations && (
-                            <StartReplayButton />
-                          )}
-                          {isExternal && (
-                            <ChatExternalControls
-                              conversations={selectedConversations}
-                              showScrollDownButton={showScrollDownButton}
-                              onScrollDownClick={handleScrollDown}
-                            />
-                          )}
-                        </ChatInput>
-                      )}
-
-                      {isPlayback && (
-                        <PlaybackControls
-                          nextMessageBoxRef={nextMessageBoxRef}
-                          showScrollDownButton={showScrollDownButton}
-                          onScrollDownClick={handleScrollDown}
-                          onResize={onChatInputResize}
-                        />
-                      )}
-                    </>
-                  )}
-                </div>
-              </div>
-              {isShowChatSettings && (
-                <div
-                  className={classNames(
-                    'absolute left-0 top-0 grid size-full',
-                    selectedConversations.length === 1
-                      ? 'grid-cols-1'
-                      : 'grid-cols-2',
-                  )}
-                >
-                  {selectedConversations.map((conv, index) => (
-                    <ChatSettings
-                      key={conv.id}
-                      conversation={conv}
-                      modelId={conv.model.id}
-                      prompts={prompts}
-                      addons={addons}
-                      onChangeSettings={(args) => {
-                        handleTemporarySettingsSave(conv, args);
-                      }}
-                      onApplySettings={handleApplyChatSettings}
-                      onClose={() => setIsShowChatSettings(false)}
-                      isOpen={isShowChatSettings}
-                      isRight={index === 1}
-                      isCompareMode={isCompareMode}
-                    />
-                  ))}
-                </div>
-              )}
-              {isCompareMode && selectedConversations.length < 2 && (
-                <div className="flex h-full w-[50%] flex-col overflow-auto">
-                  <ChatCompareSelect
-                    conversations={conversations}
-                    selectedConversations={selectedConversations}
-                    onConversationSelect={(conversation) => {
-                      dispatch(
-                        ConversationsActions.selectForCompare(conversation),
-                      );
-                    }}
-                  />
-                  <div
-                    className="shrink-0"
-                    style={{ height: inputHeight + 56 }}
-                  />
-                </div>
-              )}
-            </div>
-          </div>
-        </>
-      )}
-    </div>
-  );
-});
-
-ChatView.displayName = 'ChatView';
 
 export function Chat() {
   const { t } = useTranslation(Translation.Chat);
@@ -914,5 +63,6 @@ export function Chat() {
       />
     );
   }
+
   return <ChatView />;
 }
